Changes to the dashboard, navbar and incoming gcn notices (#31)
This PR removes the default SkyPortal dashboard to replace it with a custom grandma dashboard, changes the labels of certain components and pages to make the app easier to use during the next campaign, and adds the possibility for user to label sources: kilonova, not kilonova, gw, not gw, grb and not grb. The label shows on the recent sources widget of the dashboard. In this PR, we also remove Fermi GCN notices in favor of SWIFT notices.